Module joinkit::util

This module contains various utilities/helper functions

Functions

extract_key
Extracts a key from the record.

extract_key_value
Extracts a key from the record and returns a tuple of the key and the record.

fields_to_idx
Parses a string containing the fields indices in base1 and tries to convert them into vector of 2-element tuples.

num_fields
Returns a number of fields in the record.

rec_sep_as_byte
Converts a record separator to a single byte

write_both
Writes both, the left value and the right value into output stream.

write_left
Writes only the left value with padded field separators in place of missing right value.

write_right
Writes only the right value with padded field separators in place of missing left value.