Function joinkit::util::fields_to_idx

pub fn fields_to_idx(f: &str) -> Result<Vec<(usize, isize)>, ClapError>

Parses a string containing the fields indices in base1 and tries to convert them into vector of 2-element tuples.

Each tuple contains the parsed field index and its position, both in base0. The association with position is necessary, since the field indices will be sorted (to facilitate the key extraction in extract_key() function) and thus might loose the info about their correct position.

The resulting vector is sorted on the field indices. The error is returned if the input string contains duplicate fieled indices.

Example

use joinkit::util;

// does not need to be ordered
let field_str = "1,3,6,4";
let field_idx = util::fields_to_idx(field_str).unwrap();

// this reads as follows: the first field goes to the first position, the third field goes to
// the second position, the fourth field goes to the fourth position and the sixth field goes
// to the third position.
assert_eq!(vec![(0, 0), (2, 1), (3, 3), (5, 2)], field_idx);